requirements.txt
openai
ipykernel
langchain-community
python-dotenv
bs4
gradio
opencv-python
gtts
pillow
torch